refactor(#3479): extract turn_bridge cancel/finalize decision helpers to submodule
#3479 Phase-1 rank-3 behavior-preserving extraction. Move the pure,
unit-tested cancel/finalize-policy decision helpers out of the frozen
turn_bridge/mod.rs giant into a new sibling cancel_finalize_policy.rs,
re-exported so every call site stays byte-identical.

Moved (bodies verified byte-identical; only visibility changed to
pub(in crate::services::discord) for the re-export):
- classify_turn_finished_dispatch_kind
- is_done_setting_terminal_frame
- should_finalize_cancel_after_recv
- should_suppress_headless_delivery_for_cancel
- should_record_final_turn_transcript
- resolve_bridge_owner_channel
plus their #[cfg(test)] modules (dispatch_kind_tests,
transcript_delivery_gate_tests, cancel_recv_toctou_tests,
bridge_owner_channel_resolution_tests).

turn_bridge/mod.rs: 6811 -> 6702 prod LoC (ratchet lowered to lock the
win). New module is 131 prod + 450 test LoC (not a giant). Zero logic
change. Left behind for coupling: done_result_requires_full_terminal_replay
(its test module is entangled with the non-pure sync_response_delivery_state),
sync_inflight_restart_mode_from_cancel (mutates &mut state), and
bridge_should_reclaim_relay_from_missing_watcher (untested).
