+++ ARCHITECTURE.md
@@ -509,6 +509,7 @@ src/
 │   │   │   ├── mod.rs
 │   │   │   └── store.rs
 │   │   ├── turn_bridge/
+│   │   │   ├── cancel_finalize_policy.rs
 │   │   │   ├── completion_guard.rs
 │   │   │   ├── context_window.rs
 │   │   │   ├── headless_delivery.rs
